+++ lib/parallel_tests/grouper.rb
@@ -48,8 +48,15 @@ module ParallelTests
 
       private
 
+      def specified_groups(options)
+        groups = options[:specify_groups]
+        return groups if groups != '-'
+
+        $stdin.read.chomp
+      end
+
       def specify_groups(items, num_groups, options, groups)
-        specify_test_process_groups = options[:specify_groups].split('|')
+        specify_test_process_groups = specified_groups(options).split('|')
         if specify_test_process_groups.count > num_groups
           raise 'Number of processes separated by pipe must be less than or equal to the total number of processes'
         end